Implementing Endpoint Handlers
These handlers can help breakup the main file, by having each /api/ROOT be a different handler.

Surprisingly in testing this actual reduces response time. Not by much only about a 1ms shaved off request time across the board. But it is effective, could be due to garbage collection with separate files? Regardless a benefit, and one that makes the code easier to read and maintain