Numerous bug fixes, modifications
All of these changes amount to being able to publish new packages.

GitHub is contacted with Pulsar-Bot to find all relevant files from the repo, then will create a proper Server Package Object file, and write it to disk, as well as a proper package_pointer entry. After everything returning it to the user.